refactor: extract function compress_messages to the util
